backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic>=2.11.9
python-dotenv==1.0.0
supabase>=2.20.0